add caldavManulapi & caldavCron modulation
